% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{scalar.dist}
\alias{scalar.dist}
\alias{scalar.dist,vector-method}
\title{Calculate scalar distance}
\usage{
scalar.dist(x, ...)
}
\arguments{
\item{x}{A vector of numeric values.}

\item{\dots}{Additional arguments, not used at this time.}
}
\value{
a matrix object
}
\description{
Calculate the scalar distance between elements of a matrix.
}
\details{
Take the absolute difference between all elements in a vector, and return a
matrix of the distances.
}
\examples{

scalar.dist(1:10)

}
\author{
Jake Bowers, \url{http://jakebowers.org/}, Cole Beck
}

