% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbpmatching-package.R
\name{nbpMatching-internal}
\alias{.requireCachedGenerics}
\alias{[<-,distancematrix,ANY,ANY,ANY-method}
\alias{[<-,distancematrix-method}
\alias{[[<-,distancematrix,ANY,ANY,ANY-method}
\alias{[[<-,distancematrix-method}
\alias{full.qom}
\alias{full.qom,data.frame-method}
\alias{initialize,distancematrix-method}
\alias{nbpMatching-internal}
\alias{runner}
\alias{runner,data.frame-method}
\alias{show,nonbimatch-method}
\title{Internal nbpMatching objects.}
\description{
Internal nbpMatching objects.
}
\details{
This function should not be called by the user.
}
\keyword{internal}

