% Generated by roxygen2 (4.0.2): do not edit by hand
\name{nbpMatching-internal}
\alias{.requireCachedGenerics}
\alias{[<-,distancematrix-method}
\alias{[[<-,distancematrix-method}
\alias{full.qom}
\alias{full.qom,data.frame-method}
\alias{initialize,distancematrix-method}
\alias{nbpMatching-internal}
\alias{runner}
\alias{runner,data.frame-method}
\title{Internal nbpMatching objects.}
\description{
Internal nbpMatching objects.
}
\details{
This function should not be called by the user.
}
\keyword{internal}

