\name{nbpMatching-internal}
\title{Internal nbpMatching objects}
\alias{.First.lib}
\alias{.requireCachedGenerics}
\alias{initialize,distancematrix-method}
\alias{[<-,distancematrix-method}
\alias{[[<-,distancematrix-method}
\alias{runner}
\alias{runner,data.frame-method}
\alias{full.qom}
\alias{full.qom,data.frame-method}
\description{Internal nbpMatching objects.}
\details{This function should not to be called by the user.}
\keyword{internal}
